requirements.txt
bs4
lxml
requests
linkpreview
python-docx